+++ changelog.md
@@ -2,6 +2,22 @@
 
 ## [Unreleased]
 
+## [0.6.0] - 2023-10-18
+
+### Added
+
+- Migrated public functions to separate files.
+- Added missing online help links
+
+### Changed
+
+- Renamed `Get-BitlyGroupLinks` to `Get-BitGroupLink`. _This is a breaking change_.
+- Updated `Save-BitlyToken` to let the user pass a token as a secure string as a parameter value or from the pipeline. The default behavior is to prompt the user with a `Read-Host` command so they can type or paste the API key.
+- Revised `Get-BitlyGroupLinks` to sort output on data by default.
+- Updated help documentation.
+- Updated `README.md`.
+- Updated Pester tests.
+
 ## [0.5.0] - 2023-09-30
 
 - added `about_psbitly` help topic
@@ -37,4 +53,5 @@
 
 - initial files
 
-[Unreleased]: https://github.com/jdhitsolutions/PSBitly/compare/v0.5.0..HEAD
+[Unreleased]: https://github.com/jdhitsolutions/PSBitly/compare/v0.6.0..HEAD
+[0.6.0]: https://github.com/jdhitsolutions/PSBitly/compare/v0.5.0..v0.6.0